+++ annotations/src/main/java/com/squareup/anvil/annotations/internal/InternalAnvilHintMarker.kt
@@ -1,0 +1,7 @@
+package com.squareup.anvil.annotations.internal
+
+/**
+ * Indicates the annotated property is an anvil hint, used for in-round processing.
+ */
+@Target(AnnotationTarget.PROPERTY)
+public annotation class InternalAnvilHintMarker