+++ annotations/src/main/java/com/squareup/anvil/annotations/internal/InternalMergedTypeMarker.kt
@@ -3,7 +3,7 @@ package com.squareup.anvil.annotations.internal
 import kotlin.reflect.KClass
 
 /**
- * Metadata bout the origin of a merged type. Useful for testing and can be discarded in production.
+ * Metadata about the origin of a merged type. Useful for testing and can be discarded in production.
  */
 @Target(AnnotationTarget.CLASS)
 public annotation class InternalMergedTypeMarker(