+++ compiler/src/main/java/com/squareup/anvil/compiler/codegen/KspContributesSubcomponentHandlerSymbolProcessor.kt
@@ -1,5 +1,6 @@
 package com.squareup.anvil.compiler.codegen
 
+import com.google.devtools.ksp.getClassDeclarationByName
 import com.google.devtools.ksp.isAbstract
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.processing.SymbolProcessorEnvironment
@@ -8,7 +9,6 @@ import com.google.devtools.ksp.symbol.KSAnnotation
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.google.devtools.ksp.symbol.KSName
-import com.google.devtools.ksp.symbol.KSType
 import com.squareup.anvil.annotations.ContributesSubcomponent
 import com.squareup.anvil.annotations.MergeSubcomponent
 import com.squareup.anvil.annotations.internal.InternalContributedSubcomponentMarker
@@ -21,7 +21,6 @@ import com.squareup.anvil.compiler.PARENT_COMPONENT
 import com.squareup.anvil.compiler.codegen.ksp.AnvilSymbolProcessor
 import com.squareup.anvil.compiler.codegen.ksp.KspAnvilException
 import com.squareup.anvil.compiler.codegen.ksp.classId
-import com.squareup.anvil.compiler.codegen.ksp.contextualToClassName
 import com.squareup.anvil.compiler.codegen.ksp.declaringClass
 import com.squareup.anvil.compiler.codegen.ksp.exclude
 import com.squareup.anvil.compiler.codegen.ksp.find
@@ -33,19 +32,22 @@ import com.squareup.anvil.compiler.codegen.ksp.isInterface
 import com.squareup.anvil.compiler.codegen.ksp.modules
 import com.squareup.anvil.compiler.codegen.ksp.parentScope
 import com.squareup.anvil.compiler.codegen.ksp.replaces
+import com.squareup.anvil.compiler.codegen.ksp.requireClassDeclaration
 import com.squareup.anvil.compiler.codegen.ksp.resolvableAnnotations
 import com.squareup.anvil.compiler.codegen.ksp.resolveKSClassDeclaration
 import com.squareup.anvil.compiler.codegen.ksp.returnTypeOrNull
-import com.squareup.anvil.compiler.codegen.ksp.scope
+import com.squareup.anvil.compiler.codegen.ksp.scopeClassName
 import com.squareup.anvil.compiler.codegen.ksp.trace
 import com.squareup.anvil.compiler.contributesSubcomponentFqName
 import com.squareup.anvil.compiler.daggerBindingModuleSpec
 import com.squareup.anvil.compiler.defaultParentComponentFunctionName
+import com.squareup.anvil.compiler.fqName
 import com.squareup.anvil.compiler.internal.asClassName
 import com.squareup.anvil.compiler.internal.createAnvilSpec
 import com.squareup.anvil.compiler.internal.joinSimpleNamesAndTruncate
 import com.squareup.anvil.compiler.internal.reference.asClassId
 import com.squareup.anvil.compiler.internal.safePackageString
+import com.squareup.anvil.compiler.mapToSet
 import com.squareup.anvil.compiler.mergeComponentFqName
 import com.squareup.anvil.compiler.mergeInterfacesFqName
 import com.squareup.anvil.compiler.mergeSubcomponentFqName
@@ -169,7 +171,7 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
                 .builder(MergeSubcomponent::class)
                 .addMember(
                   "scope = %T::class",
-                  contribution.scope.contextualToClassName(contribution.annotation),
+                  contribution.scope,
                 )
                 .apply {
                   fun addClassArrayMember(
@@ -202,7 +204,7 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
                   resolver,
                   contribution.clazz,
                   factoryClass?.originalReference,
-                  contribution.parentScopeType,
+                  contribution.parentScope,
                 )
               addAnnotation(
                 AnnotationSpec.builder(InternalContributedSubcomponentMarker::class)
@@ -289,7 +291,12 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
             .flatMap { annotatedClass ->
               annotatedClass.find(generationTrigger.asString())
                 .map { annotation ->
-                  Trigger(annotatedClass, annotation.scope(), annotation.exclude().toSet())
+                  Trigger(
+                    annotatedClass, annotation.scopeClassName(),
+                    annotation.exclude().mapToSet {
+                      it.toClassName()
+                    },
+                  )
                 }
             }
         }
@@ -297,7 +304,7 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
 
     if (triggers.isNotEmpty() && !hasComputedInitialContributions) {
       hasComputedInitialContributions = true
-      populateInitialContributions()
+      populateInitialContributions(resolver)
     }
 
     // Find new contributed subcomponents in this module. If there's a trigger for them, then we
@@ -332,7 +339,7 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
       .flatMap { contribution ->
         triggers
           .filter { trigger ->
-            trigger.scope == contribution.parentScopeType && contribution.clazz !in trigger.exclusions
+            trigger.scope == contribution.parentScope && contribution.classClassName !in trigger.exclusions
           }
           .map { trigger ->
             GenerateCodeEvent(trigger, contribution)
@@ -431,7 +438,7 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
 
   private fun checkReplacedSubcomponentWasNotAlreadyGenerated(
     contributedReference: KSClassDeclaration,
-    replacedReferences: Collection<ClassName>,
+    replacedReferences: Set<ClassName>,
   ) {
     replacedReferences.forEach { replacedReference ->
       if (processedContributionClasses.any { it == replacedReference }) {
@@ -445,7 +452,9 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
     }
   }
 
-  private fun populateInitialContributions() = trace("Populate initial contributions") {
+  private fun populateInitialContributions(resolver: Resolver) = trace(
+    "Populate initial contributions",
+  ) {
     // Find all contributed subcomponents from precompiled dependencies and generate the
     // necessary code eventually if there's a trigger.
     contributions += classScanner
@@ -453,7 +462,9 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
         annotation = contributesSubcomponentFqName,
         scope = null,
       )
-      .map { clazz ->
+      .map { contribution ->
+        // TODO can we push up this data into ContributedType?
+        val clazz = resolver.requireClassDeclaration(contribution.className, node = null)
         Contribution(
           annotation = clazz.resolvableAnnotations.single { it.fqName == contributesSubcomponentFqName },
         )
@@ -483,8 +494,8 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
 
   private class Trigger(
     val clazz: KSClassDeclaration,
-    val scope: KSType,
-    val exclusions: Set<KSClassDeclaration>,
+    val scope: ClassName,
+    val exclusions: Set<ClassName>,
   ) {
 
     val clazzFqName = clazz.fqName
@@ -515,11 +526,11 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
   private class Contribution(val annotation: KSAnnotation) {
     val clazz = annotation.declaringClass
     val classClassName = clazz.toClassName()
-    val scope = annotation.scope()
-    val parentScopeType = annotation.parentScope().asType(emptyList())
+    val scope = annotation.scopeClassName()
+    val parentScope = annotation.parentScope().toClassName()
 
     override fun toString(): String {
-      return "Contribution(class=$classClassName, scope=$scope, parentScope=$parentScopeType)"
+      return "Contribution(class=$classClassName, scope=$scope, parentScope=$parentScope)"
     }
 
     override fun equals(other: Any?): Boolean {
@@ -529,7 +540,7 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
       other as Contribution
 
       if (scope != other.scope) return false
-      if (parentScopeType != other.parentScopeType) return false
+      if (parentScope != other.parentScope) return false
       if (classClassName != other.classClassName) return false
 
       return true
@@ -537,7 +548,7 @@ internal class KspContributesSubcomponentHandlerSymbolProcessor(
 
     override fun hashCode(): Int {
       var result = scope.hashCode()
-      result = 31 * result + parentScopeType.hashCode()
+      result = 31 * result + parentScope.hashCode()
       result = 31 * result + classClassName.hashCode()
       return result
     }